move nn out of symmetry.py, refactor symmetry pooling
